Add Fold/Unfold blocks to IR. Refactor IR sanity checks.
