+++ graphql_compiler/compiler/ir_lowering_gremlin.py
@@ -15,8 +15,8 @@ from ..exceptions import GraphQLCompilationError
 from .blocks import Backtrack, CoerceType, Filter, Traverse
 from .expressions import BinaryComposition, Literal, LocalField, NullLiteral
 from .ir_lowering_common import (lower_context_field_existence, merge_consecutive_filter_clauses,
-                                 optimize_boolean_expression_comparisons,
-                                 sanity_check_ir_blocks_from_frontend)
+                                 optimize_boolean_expression_comparisons)
+from .ir_sanity_checks import sanity_check_ir_blocks_from_frontend
 
 
 ##################################
